Safari-611.1.10/LayoutTests/js/dom/inc-const-valueOf-expected.txt

Test for regression against

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testPostIncConstVarWithIgnoredResult() threw exception TypeError: Attempted to assign to readonly property..
PASS testPreIncConstVarWithIgnoredResult() threw exception TypeError: Attempted to assign to readonly property..
PASS testPreIncConstVarWithAssign() instanceof TypeError is true
PASS successfullyParsed is true

TEST COMPLETE
